[IMP] knowledge: owlification TopBar
Purpose
=======

We need to transition the Knowledge app to OWL in order to remove the
renderingContext which should not have been necessary in the first place,
as per the new specs of the JS Framework.

Specs
=====

In this commit are adding a new Component called `TopBar`.
This Component uses its environment to pass all the actions to its parent:
`KnowledgeArticleFormRenderer`, in which we store all the essential methods
needed to communicate with the database.
This way, when we call a method inside the TopBar, we are directly interacting
with KnowledgeArticleFormRenderer and all the changes that needs to be saved to
the database will be done.

task-3095342

Part-of: odoo/enterprise#35261