[FIX] web_studio: kanban editor, grouped, no record
When a view is grouped by a date or datetime fields, the Group
datapoints expect to receive a `range` param (e.g. indicating that
the group corresponds to the period from 12/1/2022 to 12/31/2022).
In the KanbanEditor, when there's no data, we manually create
datapoints to simulate that there's a record (and we keep the fact
that the view is grouped). So, we manually create a Group datapoint.

Before this commit, we didn't pass the `range` param to that Group
datapoint, and it thus crashed. This commit passes an empty `range`
which does the trick.

closes odoo/enterprise#34856

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>