[IMP] stock_barcode: unfold tracked products
In the Barcode inventory Adjustment, the grouped lines for tracked
products will be unfold when selected.
The justification for this change is, when doing an inventory adjustment
including a product tracked by lots, the operator can be tempted to scan
multiple times the same lot regardless the real inventory.
By displaying sublines by default, they will see which lots are supposed
to be in the inventory.

task-3036810

closes odoo/enterprise#33179

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>