[FIX] web_studio: kanban dropdown defined only by o_kanban_manage_button_section is clickable
Have a kanban arch that defines its dropdown only with the class "o_kanban_manage_button_section".
Enter studio, click on the dropdown ellipsis icon.

Specifically, in the App Helpdesk on the "Helpdesk Overview" kanban, enter studio. The crash should happen
if this revision is not applied.

Before this commit, there was a crash as the xpath of the dropdown was not set.

After this commit, the dropdown is fully recognized as such.

closes odoo/enterprise#34891

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>