[IMP] account_reports: Improve the account filter on reports
Currently, the account filter present in the reports works well with
small size report but on big ones (e.g. big partner ledger reports) the
filtering can take a very long time due to too many jquery selector
calls. So for every key stroke, the report takes an important amount of
time to finish to function, which also makes for a poor user experience (e.g.
when someone type multiple strokes, the filter bar doesn't show the
characters).

The function was changed to reduce these calls. By doing so we went from
~4.3 seconds to 243 milliseconds for partner ledger with nearly 5.500
lines when filtering with one character.

This PR is related to a ticket in 15.0, but the problems is also present
in 14.0.

closes odoo/enterprise#36189

X-original-commit: 0f417a9e8193c6c2f7aabc81161810bfcc0c1b3f
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Simon Morais Da Cunha (smdc) <smdc@odoo.com>