[FIX] mrp_workorder: refresh quantity of material
In the new version 16.0, changing the quantity produced on the tablet view of a workorder does not automatically change the quantity of consumed material when focusing out of the input field.
The only way to do so is to do another action (e.g. press the menu button).
This patch fixes this issue. How it works:

- in tablet.js, if there are steps in the workorder, we create the onRecordChanged function which will save the workorder record and re-render if there's a change on the workorder
- in form_view.js, we override the update method to include the function onRecordChanged passed into props
- in mrp_workorder.py, we make sure that it impacts all checks where a product component is registered

closes odoo/enterprise#35713

X-original-commit: 8962a476154b94f8046a4bc5a0ae3d4a131978c4
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Diez Stéphane (snd) <snd@odoo.com>