[REF] sale_subscription: factorize payment values logic in controllers
The payment values in the Qweb context of sale, sale_subscription and
website_sale were computed separately, although there is a large common
basis. This split made it hard to communicate between modules and
generated a lot of duplicates. The new function `_get_payment_values`
in sales is now used as a common basis method for all sale_* modules to
get the common payment values.

closes odoo/enterprise#34922

Related: odoo/odoo#107788
Signed-off-by: Vallaeys Valentin (vava) <vava@odoo.com>