[FIX] appointment_hr: get workhours in multi-company websites
Current behaviour:
If a user that belongs to Company A, has defined workhours in Company A,
generates an invite link based on workhours, when you open the link,
on a website that belongs to another company, the appointment slots are
ignoring the workhours of the employee.

Expected behaviour:
A portal user shouldn't be able to book appointments outside workhours,
when that invite link was based on the employee's workhours,
regardless of the company the sites belongs to.

Steps to reproduce:
- Install Employees, Website, Appointment
- Create a new company B
- In Settings, for website 2, set it's company to company B
- In Calendar, generate a new invite link based on workhours
- Go to the link and preview it under website 2,
you will see that you can book slots outside the normal 40h/workweek
default schedule.

Reason for the problem:
When calculating the `_slot_availability_prepare_values_workhours`,
the `allowed_company_ids` in the context of `staff_users` is that of the website's company,
which means that we get an empty record when getting the `employee_id`.
Therefor we can't get the calendar resource of the user, which means we fallback
on the generic slots, instead of the workhours.

Fix:
When computing the workhours, if the user doesn't have an employee,
we look at all his employees regardless of the company and we take the first one.

Affected versions:
- saas-15.2
- saas-15.3
- 16.0
- saas-16.1
- master

opw-3087872

closes odoo/enterprise#39040

X-original-commit: e6cd8b0b79b33fdd1548c9794913b207d63b2319
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>