[IMP] l10n_be_reports: add official BS and P&L versions
Currently, the BS and P&L version for Belgium was a combination of
different old versions and not representing the latest official
versions.

In Belgium there are 4 BS versions for companies, depending on the size
of the company and the company form. For associations and foundations
there are 2 BS versions, depending on the size.

For the P&L we have 2 version for companies and 2 versions for
associations and foundations, both depending on their size.

All versions have been implemented and will be available based on the
selected CoA (either for companies or associations and foundations).

task-3086473

closes odoo/enterprise#35771

Related: odoo/odoo#117480
Related: odoo/upgrade#5291
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>