[FIX] account_consolidation: correcting wrong required attribute condition
install account_consolidation module and try creating record in Consolidation -> Configuration -> Company periods, in the form currently the Period will not have required attribute and in the model level the field is required.

So if the user tries to create a record without inputting  the value to the period field, the exception will be shown.

this pr will change the given required attribute from view level, as it is already required at model level

closes odoo/enterprise#35863

X-original-commit: b8c98bf51fb2df94053e7e8c593cb0b920a62d82
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>