[IMP] account_consolidation: form alignment
install account_consolidation module and navigate to Consolidation -> Configuration -> Periods, open any existing record or create new record and click on the smart button named Edit.

from the next screen click on Add a Column button, in the opening form view, the two fields will be visible and other fields are invisible. as other fields are invisible and when the two fields align in two groups at different level. moving the field to first group for better ui

closes odoo/enterprise#34947

Signed-off-by: Laurent Smet <las@odoo.com>