[FIX] account_reports: Unbalanced trial balance
Before this commit, the trial balance was not balanced when using a period other that a fiscal year.

The problem was due to an error in the date_to computation of the initial balance when the key 'include_current_year_in_unaff_earnings' was True.

We were computing the initial balance date_to once in '_get_options_initial_balance' and then again in '_get_options_unaffected_earnings' when we should just use the date_to already computed in '_get_options_initial_balance'.

closes odoo/enterprise#35672

Task-id: 3099868
X-original-commit: 5ef5e2e1ad87dc6a8016c7dc982254ad410e3189
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: rugo-odoo <rugo@odoo.com>