[FIX] sale_susbcription: upsell improvement & fix
This commit aims to allow upsell on renewal to be correctly computed as well as simplify the discount logic.
We also allow to upsell a SO that has not been invoiced yet.

Before this commit, the prorata was based on the number of day in the current month, this could lead to some weird behaviour where 5 days were not treated the same dependins on when we were upselling.

Now, we do a prorata based on the average number of day in a month to ensure a more coherent results.

task-id : 3086493

closes odoo/enterprise#39109

Forward-port-of: #35711
X-original-commit: 2e98523c5e29de257551258a4affe8b0f32ac41c
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Damhaut Florian (flda) <flda@odoo.com>