[IMP] account_disallowed_expenses_fleet: use smallest rate
At this point of time in Odoo, if 2 different rates are set
on a Purchase line (one via the Account and the DNA category,
a second found via the Vehicle), the rate coming from
the vehicle takes precedence.

In the context of the regulation change, we would like to be
more precise than that, and pick the smallest of the two rates.

task-3082174

closes odoo/enterprise#34995

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>