[FIX] sale_subscription: remove invoice trigger trigger for new invoices
Before this commit, the subscription invoice creation was triggered each
time a new invoice was ready to be sent. it occured when the invoice
cron was triggered. As the two cron would try to modify the same sale.order
concurrent update would be triggered regularly.

This change was introduced by the following commit:
https://github.com/odoo/enterprise/commit/59d8806e8da0c3e89cfc14d3efa5bb9ade3f1095

Taskid: 3054994
Part-of: odoo/enterprise#35710