[IMP] account_reports: Partner Ledger, Aged Receivable and Aged Payable: improve performance when too many lines
Those 3 reports became unusable when they contained too many lines. This happened when a big number of partners were involved in the transactions they reported.

When opened on Chrome, scrolling was only possible with big saccades ; on Firefox, the scrolling was smooth but the report took way longer to open. This was due to the style rendering. On Chrome, it is done in a lazy manner, directly when scrolling ; Firefox preloads everything.

On even larger number of lines, the reports even failed to open, raising a MemoryError during the Qweb rendering of the html.

To solve that, we choose to give the option to regroup the lines of those report by prefix of their name. The first character of each line will be used to create an index in the UI: for examples, 'Arnold' and 'Arthur' would both be grouped under a common line 'A'. By default, we only display those prefix lines, tremendously reducing the number of lines to be rendered in the report itself. Those lines can then be unfolded manually by the user depending what he wants to see. If unfolding a line should inject too many lines into the report again (for example, thousands of partner names could start with letter 'A'), the prefix grouping mechanism is called recursively. We will then generate under line A sublines AA, AB, AC, ... (of course not generating lines for prefixes not matching any partner). When unfolding a prefix line, if only a reasonable number of lines should be loaded, we don't generate sub-prefixes, but directly add all the partner lines matching the parent prefix. The report structure hence adapts itself to the data, keeping it comfortable for the user.

In order to make this work, some refactoring was also necessary on the way the totals below sections were generated, in order to standardize their behavior on groupby lines.

This new prefix grouping feature is by default NOT activated. It can be enabled separately for Aged Payable/Receivable and Partner Ledger by defining new config parameters, and providing as its value the threshold defining what we consider as "too many sublines". When the number of partner lines to display is greater than that threshold, prefix groups are used automatically. Here are the names of these config parameters:
- For AP/AR: account_reports.aged_partner_balance.groupby_prefix_groups_threshold
- For Partner Ledger: account_reports.partner_ledger.groupby_prefix_groups_threshold

Solves issue https://github.com/odoo/odoo/issues/77529

Part-of: odoo/enterprise#37484