[FIX] account_saft: fix validation errors
In Odoo, it is currently required for every customer/vendor to have a mobile or phone number in order to export the SAF-T report. However, in reality (documentation), this field is only required for the header company, and optional for customers/vendors.

The SAF-T report needs contacts information. Currently, every type of partner is picked for contacts, while it would make more sense to only pick the 'contact' type. This would also solve an issue with partners' name being empty (which can be the case for partners with a type different from 'contact', like Invoice or Delivery addresses).

This commit makes the Telephone field optional, and modifies the error raised to only check for the phone of the company. It also filters contacts on the 'contact' type to avoid errors for empty contact names.

task id=3099751

closes odoo/enterprise#35338

X-original-commit: daf674c3a81fbcba3485a083549e45d6ad26cec7
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Popeler Antoine (anpp) <anpp@odoo.com>