[FIX] knowledge: fixes crash on deleted nested articles
This commit fixes an issue where the /outline command would crash if
we are trying to access an article that was previously deleted.

Now we check if the article needs to be deleted, if it was done
then this raises a user Error informing the user that the article
doesn't exist anymore.

task-3101171

closes odoo/enterprise#34985

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>