[FIX] account_reports_cash_basis: use latest date when an invoice is reconciled with an older payment
Before this fix, when an invoice was reconciled with a payment made on a prior date, the invoice appeared in the report at the date of the payment. This was wrong: in this case, we want to use the most recent date: if the payment is in the past, the invoice date must be used.

X-original-commit: f3346c12e36b1d19decf610d24453afeace3e42b
Manual forward-port of https://github.com/odoo/enterprise/pull/33007

closes odoo/enterprise#35892

Signed-off-by: Laurent Smet <las@odoo.com>