[IMP] l10n_mx_edi_website_sale: add cfdi_to_public field
Allow to send a CFDI as "publico en general" by adding a field on the
account.move and on the sale.order. Also add the field
`l10n_mx_edi_usage` on the SO. When creating a move from a SO, the
values of these fields are copied.

In addition, add an extra tab "Invoicing Info" on the ecommerce for the
customer to indicate whether he wants an invoice or not. If he wants an
invoice, ask for the fiscal regime, the usage and the tax breakdown. If
he doesn't want an invoice, it means we can send the related invoice in
"publico en general" mode.

The extra info tab appears only if:
* the company is mexican
* the automatic_invoice setting is True

When displaying the extra info tab, we also need to add a flag to all
the templates rendered during the checkout flow (because each of them
calls the wizard_checkout template, responsible for the rendering of a
continuous line on top of the screen with the current step of the
checkout we are at), these template's ids are:
* cart
* address
* checkout
* extra_info
* payment

task-3089169

closes odoo/enterprise#35050

Related: odoo/odoo#108181
Signed-off-by: Laurent Smet <las@odoo.com>