[IMP] helpdesk(_sale_timesheet): improve UX of SLA form
As part of this task,  add muted text above the team id field, added muted
text in muted on the right side of the tag_ids field. Display an "OR" in a muted
on right of the partner_ids field and shift the sale_line_ids field next to it.

task-3102422

Part-of: odoo/enterprise#35184