[FIX] stock_barcode: add print commands
It is currently not possible to print the delivery slip/the picking
operation thanks to the commands

To reproduce the issue:
1. Create a storable product P:
   - Barcode: 123
2. Open the Barcode app
3. Print the commands for Inventory
   - As you can see, there is the possibility to print the picking
     operation (barcode value: O-BTN.print-op) and the delivery slip
     (barcode value: O-BTN.print-slip)
4. Barcode > Operations > Receipt, Create
5. Scan 123
6. Scan O-BTN.print-slip

Error: We expect the delivery slip to be printed. Instead, there is
an error message "You are expected to scan one or more products."

There is nothing in the code that links the commands `O-BTN.
print-op` and `O-BTN.print-slip` to the appropriate actions. However,
there is the possibility to print both PDF manually (via the gear
button) since there is a print method on the component `MainComponent`

OPW-3075442

closes odoo/enterprise#35112

X-original-commit: 851c1b3a9ad761424d1b4a3a04e4b4ade50fd89e
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>