[FIX] sale_subscription: correct flow to delete subscriptions
Steps to reproduce:
- install "sale_subscription" module;
- create a subscription;
- try to delete it;

Issue:
We recommend to cancel the subscription before deleting it.
But we only have the possibility of closing it.

Cause:
There is no method called to handle a `UserError`
when unlinking the "`sale_subscription`" record.

Solution:
Add a method with @api.ondelete which overrides `_unlink_except_draft_or_cancel` method
to show a consistent error message to the client when trying to delete a subscription.
Add the possibility to cancel a subscription if it is closed.

opw-3087387

closes odoo/enterprise#35561

X-original-commit: ce93686d797c1458f55cb0b4a932b526c59dfdce
Signed-off-by: Arnaud Joset <arj@odoo.com>