[FIX] purchase: added price_unit to _can_be_paid dependence
What is the current behavior that you observe?

If we change the quantity then the "should be paid" field is set to "exception". But if we change the price it stays "yes".

What would be your expected behavior in this case?

According to the documentation, "if you edit a draft bill to increase the billed quantity, change the price, or add other products, the Should Be Paid status is set to Exception." cf. https://www.odoo.com/documentation/master/applications/inventory_and_mrp/purchase/manage_deals/control_bills.html#determine-when-to-pay-with-3-way-matching

opw-3090755

closes odoo/enterprise#35470

X-original-commit: cd9a61a3d3525f3486af7c11668aa85b8d48cb97
Signed-off-by: John Laterre (jol) <jol@odoo.com>