[IMP] sale_subscription: don't compute recurring price
When multiple recurring prices are defined, the selected one is now
chosen instead of computing the lowest price.
On a renewal SO, lines are now unlinked from the parent order if the
recurrence period is changed. This behavior is added to the relevant
test.
Also, this commit includes some unrelated fixes for the same test.

task-2939221

closes odoo/enterprise#34050

Signed-off-by: Arnaud Joset <arj@odoo.com>