[FIX] {,l10n_be_}hr_payroll: move `is_credit_time` val in l10n_be
Current behaviour:
Creating work entries in non-belgian company
for an employee that is working part-time,
whos calendar schedule creates credit times fails.

Expected behaviour:
Work entries should be generated correctly

Steps to reproduce:
- Install hr_payroll, hr_work_entry_contract
- create a new resource calendar of 20 (the name is 20 hours, but just duplicate the default one
and make sure you change the Sunday instead of Friday)
- update default resource calendar (of 40hr) from Sunday - Thursday 10:00 - 18:00 hr
- Create a contract for a Test employee (which is not associated with any user)
where the working schedule is a calendar of 20hr and enable the part-time tab set the schedule to 40hr/week
- Click on the "Work Entry" smart button -> stacktrace
(Check video in ticket logs for more detailed setup)

Reason for the problem:
There is a key `is_credit_time` in the values used to create the credit times
for the contract. But that field is only defined in the `l10n_be` payroll module
localisation.

Fix:
Remove the key `is_credit_time` during creation in `hr_payroll` module
and re-add it in an override in the `l10n_be_hr_payroll` module.

Affected versions:
- 16.0
- master

opw-3088250

closes odoo/enterprise#35044

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>