[IMP] l10n_be_report: get vat export comment from wizard
Up until now, the comment in the xml exported from the Belgian tax report is taken from the report manager.
The report manager isn't really used anymore and isn't easily found/available.

This will change it so that a new comment field is available in the export wizard,  and the  value will be taken from there.

Task id #3103816

closes odoo/enterprise#35086

Signed-off-by: William André (wan) <wan@odoo.com>