[IMP] knowledge, web_gant: rename Create button to New
This commit tries to unify the name of the create buttons in different
apps. "New" is used over "Create" to avoid confusion regarding its effect.
"Create" could be interpreted as we want the current record to create a new one.
But "New" implies that the user will land on an empty page that can be filled.

task id: 3074472

closes odoo/enterprise#35139

Related: odoo/odoo#108439
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>