[FIX] hr_appraisal: smart button showing proper date for different timezones
Steps to reproduce:

- Go to the Appraisal app and create an appraisal for an employee.
- Change your timezone from users and from the browser to match, i.e
the west coast of the US.
- Create a meeting with this employe at a specific date and time.
(that will be a different day for a certain timezone), like for example
if we create a meeting for 19:00 in the west coast of the US for the 19,
it will be the 20th in Europe.
- Now check the smart button date.

Issue:

The smart button will not match the date that we have in the calendar
as it will always be computed in UTC.

Solution:

Add the timezone of the user to the date that we are computing so we get
the proper date for each user timezone.

opw-3075283

closes odoo/enterprise#35125

X-original-commit: 0ad6570f8fe922a5a9d6077281ab7b4ca1557170
Signed-off-by: Kevin Baptiste <kba@odoo.com>