[IMP] knowledge: confirm label change based on move
PURPOSE

Currently, while moving an article from one category to another, the confirmation
dialog shows the 'OK' label of the button. So it does not clarify where the
article is actually being moved.

SPECIFICATIONS

For better clarification, the confirm label(OK) will be replaced by 'MOVE TO
WORKSPACE/PRIVATE/SHARED' depending on the move.

Task-3102492

closes odoo/enterprise#35417

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>