[FIX] knowledge: user can create a copy of an article with read access
Users with "read" access on a knowledge article should be able to create a
private copy of that article using our "Create a Copy" feature.

There is nothing blocking on access-rights level, so we just adapt the attrs on
the form view to show that button to users with read access.

Task-3104706

closes odoo/enterprise#35171

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>