[FIX] web_enterprise: fix non deterministic webclient tests
Before this commit, the modified tests sometimes failed on runbot.
This is because we quickly toggled twice the home menu, and expected
the underlying view to be displayed right after the second toggle.

However, the toggle function of the home_menu service waits for an
extra tick at each toggle for the url to be updated, and the toggle
operation is done in a mutex to prevent overlapping toggles. As a
consequence, in the faulty tests, it sometimes had to wait a bit
before toggling the second time, and thus the home menu was still
displayed when we expected the underlying view to be back.

Fixing runbot errors 7120 and 4848

closes odoo/enterprise#35138

X-original-commit: 1a19e108166961724f3b9c871169ee2003e0f5b2
Signed-off-by: Géry Debongnie <ged@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>