[FIX] knowledge: template paste cannot find correct descendant element
This commit fixes a bug introduced with the encapsulation of the HTML fields
inside of a div rendering the element with class `.odoo-editor-editable` no
longer the direct children of the `.o_field_html`.

Now for the paste we are targeting `.odoo-editor-editable` as a descendant
and no longer as a direct child in the CSS Selector.

[IMP] knowledge: add test tour for useAsDescription button

This commit adds a new test tour for one of the features of Knowledge
which is to use a template as a description or message in other modules.

Since the two operations have the same flow we only test the useAsDescription
button of the templates.

This tour enables us to test if the pasting part of the operation works as
intended in the CRM module, the flow is the same in other modules that can
use Knowledge's templates.

task-3107310

closes odoo/enterprise#35132

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>