[FIX] knowledge: properly dispose of Behavior Component App
Knowledge Behaviors are actually small OWL Apps because we need to be able to
mount them on the fly in the odoo editor.

They actually were not properly destroyed when the html_field component was
destroyed, because they were not added to the set which was supposed to keep
their reference.

Simplify the intended cleaning process to destroy every created App when
destroying the html_field component, instead of when scanning for new Behaviors.
This allows to keep the App alive when it is removed when the user removes an
element in the html_field then types CTRL+Z to undo the change (which makes the
element reappear, along with the still alive Behavior OWL App).

Task-3101553

Part-of: odoo/enterprise#36232