[FIX] knowledge: use Markup on data-prop-name html_field elements
Previously, Knowledge Behavior Components were allowed to use Markup on some of
their props no matter where they came from. It is not secure as props can come
from the JSON object stored in data-behavior-props which is not sanitized. Now,
only props coming from data-prop-name (which are actually stored as Element
nodes in the html_field value, which is sanitized) will use Markup.

Modify the wysiwyg template for the /file command in order to have all data
necessary for the props of the Behavior Component in the DOM. It will be
required when we enable the collaborative mode as the nodes inserted by the
wysiwyg command are the ones that will be shared, and the rendering of the
behavior will be done independently for each collaborator.

The wysiwyg method "notifyNewBehaviors" can not be allowed to store props values
for that reason.

Task-3101553

Part-of: odoo/enterprise#36232