[FIX] knowledge: store embedded views act_window.help as DOM elements
act_window objects for embedded views are stored as a JSON string in
`data-behavior-props` attribute in the html_field value. Since the help field is
html that will be rendered if the view is empty, it is not a good idea.

Instead, the proposed solution is to put the content of the help field in the
html_field value as a hidden HTMLElement, so it could even be edited in the
future.

Currently existing embedded views won't use the help field stored in
data-behavior-props anymore (and use the default help message instead), and new
ones will use the sanitized DOM element, so there is no need for a DB "upgrade".

Task-3101553

Part-of: odoo/enterprise#36232