[FIX] stock_barcode: test_put_in_pack_scan_package
In the tour `test_put_in_pack_scan_package`, we get the line for
product1 and check its package.
Before the v16.0, only the lines for the current location were displayed
but it's not the case anymore, which means there is two lines for
product1 now when we do this verification.
To avoid issue, it's better to strengthen the selector to be sure to get
only the wanted line.

runbot build error: 12273

closes odoo/enterprise#35135

Signed-off-by: Tiffany Chang <tic@odoo.com>