[FIX] account_invoice_extract: fix PDF text unselectable in PDF
Since the Owl conversion of the boxes interface, it wasn't possible
to select text in the PDF anymore.

This was due to the box layer still being rendered even when no field
was active.
Note that there is still a limitation, the text won't be selectable
when boxes are rendered.

This fix is actually the same than what was done in commit
68e6a90c403df563f8e5862cfe4a7131a68f9cd3, but the reasons behind them
are different.

Task #3107431

closes odoo/enterprise#35137

Signed-off-by: Florian Daloze (fda) <fda@odoo.com>