[IMP] marketing_automation: handle adding child activities out of kanban arch
The addition of child activities is now done using a regular event listener
that is added in the KanbanRecord override instead of a t-on-click.

The reason for this is that the arch definition of views should be agnostic of
implementation details of its JS class.
This allows the view to work in cases where we don't spawn the JS class
(e.g: studio) and also makes it easier to extend / migrate.

Task-3070918

Part-of: odoo/enterprise#34049