[IMP] appointment: use previous duration as default for custom links
When clicking on the calendar when adding slots to the custom share
link in preparation, the default duration is always 30 minutes.
However, when selecting several slots with another preferred duration,
one must drag the slot every time.

Therefore, as long as the user is on the calendar page, the slot duration
will be by default the last one used if any, or 30 minutes.

Task-3107362

closes odoo/enterprise#35151

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>