[IMP] account_accountant: bank recon displays matching AML's first
The bank recon widget needs to display move lines that match the bank transaction value first. This will help the user find corresponding items easier.

It was decided that a custom 'order by' would have had a significant performance impact, therefore, to achieve the goal, A search for matching residual amounts is executed and the account.move.line ids are placed in the context. This context key is then used to activate a custom search read on account.move.line when the embedded tree view loads.

Task-3042547

closes odoo/enterprise#36347

X-original-commit: 15805358a0a0f8aa31c2306d6bfeb60e9bec3237
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Ayob Habib (ayh) <ayh@odoo.com>