[IMP] web_cohort, *:  improve view scale selector style and behavior
*: hr_payroll, web_gantt, web_grid

This commit improves the consistency of scale selectors in multiple views.
It is the enterprise counterpart of PR (1), which moved the calendar
selector to a new component. This component is used by the Cohort view,
already converted to Owl. Gantt and Grid views have been adapted as those
views are not yet converted.

Only the xml structure has been adapted in the legacy implementations of
those views. For the next implementation, the ViewScaleSelector component
should be used instead.

Some tests and selectors in xpath have been fixed to prevent crashes.

(1): https://github.com/odoo/odoo/pull/109090

part of task #2867569

closes odoo/enterprise#35435

Related: odoo/odoo#109090
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>