[IMP] sale_subscription: compute last invoice date based on next invoice date
Before this commit, the last_invoice-date was computed if invoices were created.
The invoice existence is depending on the link sale.order.line and account.move.line

This commit helps computing the ast_invoice_date based on SO's dates instead of
this link because in some case, real database could have missing data.
As the dates are readonly for simple salesmen, we can assume that a date difference
is enough to say that a subscription started to invoice.

Part-of: odoo/enterprise#35082