[IMP] timer: improve performance
This commit changes the search method of `is_running_timer` field in
`timer.mixin` to use the `_search` method and use the query getting
as result by that method in the domain resulted by the search method
instead of doing a search and then a mapped to get the res_ids linked
to a running timer.
This commit also uses the `_read_group` method instead of `read_group`
is a compute.

task-2702636