[IMP] social: reduce the height of the feed refresh ribbon
This reduces the feed refresh ribbon height that was too big and lighten its
hover/focus color.

Technical note: the ribbon uses a bootstrap alert which has more padding in
16.0 than in previous versions. To solve the problem, we have forced the
padding to a similar value using a bootstrap class (py-2).
Also removed the hover color change (based on color-contrast) because it was
not great in dark mode.

Task-3105171

closes odoo/enterprise#36405

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>