[IMP] l10n_be_hr_payroll: simplify double holiday pay
Prior to this commit, HR had to manage double holidays manually.
They had to configure previous occupations by putting in the relevant
lines and also removing those lines, once they were not relevant any longer.

All of this can be automated. Extra manual work and synchronization
of the occupation lines could be avoided.

After this commit it is going to be possible to directly encode the data
contained in the holiday certificates that we receive from the employee
when they starts at Odoo, and not to have a mass of manual work in June
to take over the double pay data which is on each holiday certificate
for each employee who entered into service.

We are interested in previous occupations that happened during the first
contract year of the employee and a year previous to that. That's why
we simplify the view and make visual distinction between these fields.

Also, when making double holiday pay, we compensate the year and the
previous year. Thus, we filter out automatically the occupation lines
that are no more relevant.

task - 3081167

closes odoo/enterprise#37894

Signed-off-by: Kevin Baptiste <kba@odoo.com>