[FIX] sale_temporal: fix domain for product.pricing view
The tree view for `product.pricing` seems to be copy-pasted from `product.template`
view (tree form for `product_pricing_ids` inside the `product.template` view),
because it has reference to `parent` in domain attribute, which is not available
in regular tree view.

Fix it be replacing `parent.id` to `product_template_id`.

Also, remove `column_invisible`, which doesn't make sense in this context (it's
not a tree view inside form view)

STEPS:

1) Go to Sales
2) Configuration-> Quotation templates
3) Click on "Monthly"
4) Click on "prices" button
5) Click on "Product Variants"

opw-3103960

closes odoo/enterprise#35246

Signed-off-by: Arnaud Joset <arj@odoo.com>