[FIX] knowledge: check article from MoveTo Dialog
This commit fixes a bug which caused a crash if you do not select a destination
when you want to move an article via the MoveTo dialog.

Now if we do not get any data from the input we simply return from the event
until the input contains a destination.

task-3112187

closes odoo/enterprise#35248

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>