[FIX] knowledge: check article from MoveTo Dialog
This commit fixes a bug which caused a crash if you do not select a destination
when you want to move an article via the MoveTo dialog.

Now if we do not get any data from the input we simply return from the event
until the input contains a destination.

task-3112187

closes odoo/enterprise#35292

X-original-commit: 7730a3228f80c437eeb6e383b2e986cdf61111b3
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>